Inline sub-structure headers into stinger struct
The initial implementation of the DISTRIBUTED allocation strategy
introduced an extra layer of indirection. For example, struct stinger
contained a pointer to struct stinger_vertices, which contained
a pointer to the actual array of vertices.

Now, the metadata for each sub-structure lives inside the struct
stinger itself. The corresponding init method for each sub-structure
allocates the distributed array as necessary. A new deinit function
for each structure frees the distributed array in place, without
freeing the struct itself.

With this change, the stinger struct contains all the data needed to
access any of the distributed arrays without doing a pointer
dereference. The stinger struct can be replicated to each nodelet
with a simple memcpy.